core(opportunities): more granular score
